backend/requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
pydantic-settings
openai
google-genai
